+++ CHANGELOG.md
@@ -30,6 +30,7 @@
     + `-[UIDevice modelName]`
     + `ESNetworkHelper` class
     + `-[UIView snapshotViewAfterScreenUpdates:]`
+    + `-[UITableView performBatchUpdates:]`
 - Renamed:
     + `ES_STOPWATCH_BEGIN` `ES_STOPWATCH_END` => `ESBenchmark()`
     + `ESDispatchOnMainThreadAsynchrony` => `es_dispatch_async_main`